Struct google_dfareporting2d5::Conversion

pub struct Conversion {
    pub ordinal: Option<String>,
    pub timestamp_micros: Option<String>,
    pub kind: Option<String>,
    pub child_directed_treatment: Option<bool>,
    pub encrypted_user_id: Option<String>,
    pub custom_variables: Option<Vec<CustomFloodlightVariable>>,
    pub floodlight_configuration_id: Option<String>,
    pub mobile_device_id: Option<String>,
    pub value: Option<f64>,
    pub limit_ad_tracking: Option<bool>,
    pub floodlight_activity_id: Option<String>,
    pub quantity: Option<String>,
}

A Conversion represents when a user successfully performs a desired action after seeing an ad.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

batchinsert conversions (none)

Fields
ordinal: Option<String>

The ordinal of the conversion. Use this field to control how conversions of the same user and day are de-duplicated. This is a required field.

timestamp_micros: Option<String>

The timestamp of conversion, in Unix epoch micros. This is a required field.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#conversion".

child_directed_treatment: Option<bool>

Whether the conversion was directed toward children.

encrypted_user_id: Option<String>

The alphanumeric encrypted user ID. When set, encryptionInfo should also be specified. This field is mutually exclusive with encryptedUserIdCandidates[] and mobileDeviceId. This or encryptedUserIdCandidates[] or mobileDeviceId is a required field.

custom_variables: Option<Vec<CustomFloodlightVariable>>

Custom floodlight variables.

floodlight_configuration_id: Option<String>

Floodlight Configuration ID of this conversion. This is a required field.

mobile_device_id: Option<String>

The mobile device ID. This field is mutually exclusive with encryptedUserId and encryptedUserIdCandidates[]. This or encryptedUserId or encryptedUserIdCandidates[] is a required field.

value: Option<f64>

The value of the conversion.

limit_ad_tracking: Option<bool>

Whether the user has Limit Ad Tracking set.

floodlight_activity_id: Option<String>

Floodlight Activity ID of this conversion. This is a required field.

quantity: Option<String>

The quantity of the conversion.