Struct google_dfareporting2d5::DefaultClickThroughEventTagProperties

pub struct DefaultClickThroughEventTagProperties {
    pub override_inherited_event_tag: Option<bool>,
    pub default_click_through_event_tag_id: Option<String>,
}

Properties of inheriting and overriding the default click-through event tag. A campaign may override the event tag defined at the advertiser level, and an ad may also override the campaign's setting further.

This type is not used in any activity, and only used as part of another schema.

Fields
override_inherited_event_tag: Option<bool>

Whether this entity should override the inherited default click-through event tag with its own defined value.

default_click_through_event_tag_id: Option<String>

ID of the click-through event tag to apply to all ads in this entity's scope.