Struct google_dfareporting2d5::DirectorySite

pub struct DirectorySite {
    pub kind: Option<String>,
    pub country_id: Option<String>,
    pub description: Option<String>,
    pub settings: Option<DirectorySiteSettings>,
    pub currency_id: Option<String>,
    pub inpage_tag_formats: Option<Vec<String>>,
    pub url: Option<String>,
    pub contact_assignments: Option<Vec<DirectorySiteContactAssignment>>,
    pub parent_id: Option<String>,
    pub interstitial_tag_formats: Option<Vec<String>>,
    pub id_dimension_value: Option<DimensionValue>,
    pub active: Option<bool>,
    pub id: Option<String>,
    pub name: Option<String>,
}

DirectorySites contains properties of a website from the Site Directory. Sites need to be added to an account via the Sites resource before they can be assigned to a placement.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get directory sites (response)

list directory sites (none)

insert directory sites (request|response)

Fields
kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#directorySite".

country_id: Option<String>

Country ID of this directory site.

description: Option<String>

Description of this directory site.

settings: Option<DirectorySiteSettings>

Directory site settings.

currency_id: Option<String>

Currency ID of this directory site. Possible values are: - "1" for USD - "2" for GBP - "3" for ESP - "4" for SEK - "5" for CAD - "6" for JPY - "7" for DEM - "8" for AUD - "9" for FRF - "10" for ITL - "11" for DKK - "12" for NOK - "13" for FIM - "14" for ZAR - "15" for IEP - "16" for NLG - "17" for EUR - "18" for KRW - "19" for TWD - "20" for SGD - "21" for CNY - "22" for HKD - "23" for NZD - "24" for MYR - "25" for BRL - "26" for PTE - "27" for MXP - "28" for CLP - "29" for TRY - "30" for ARS - "31" for PEN - "32" for ILS - "33" for CHF - "34" for VEF - "35" for COP - "36" for GTQ - "37" for PLN - "39" for INR - "40" for THB - "41" for IDR - "42" for CZK - "43" for RON - "44" for HUF - "45" for RUB - "46" for AED - "47" for BGN - "48" for HRK

inpage_tag_formats: Option<Vec<String>>

Tag types for regular placements.

Acceptable values are: - "STANDARD" - "IFRAME_JAVASCRIPT_INPAGE" - "INTERNAL_REDIRECT_INPAGE" - "JAVASCRIPT_INPAGE"

url: Option<String>

URL of this directory site.

contact_assignments: Option<Vec<DirectorySiteContactAssignment>>

Directory site contacts.

parent_id: Option<String>

Parent directory site ID.

interstitial_tag_formats: Option<Vec<String>>

Tag types for interstitial placements.

Acceptable values are: - "IFRAME_JAVASCRIPT_INTERSTITIAL" - "INTERNAL_REDIRECT_INTERSTITIAL" - "JAVASCRIPT_INTERSTITIAL"

id_dimension_value: Option<DimensionValue>

Dimension value for the ID of this directory site. This is a read-only, auto-generated field.

active: Option<bool>

Whether this directory site is active.

id: Option<String>

ID of this directory site. This is a read-only, auto-generated field.

name: Option<String>

Name of this directory site.