Struct google_dfareporting2d5::Placement

pub struct Placement {
    pub comment: Option<String>,
    pub campaign_id: Option<String>,
    pub payment_source: Option<String>,
    pub directory_site_id_dimension_value: Option<DimensionValue>,
    pub advertiser_id: Option<String>,
    pub key_name: Option<String>,
    pub directory_site_id: Option<String>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub content_category_id: Option<String>,
    pub size: Option<Size>,
    pub archived: Option<bool>,
    pub publisher_update_info: Option<LastModifiedInfo>,
    pub tag_formats: Option<Vec<String>>,
    pub payment_approved: Option<bool>,
    pub tag_setting: Option<TagSetting>,
    pub id: Option<String>,
    pub external_id: Option<String>,
    pub id_dimension_value: Option<DimensionValue>,
    pub account_id: Option<String>,
    pub ssl_required: Option<bool>,
    pub status: Option<String>,
    pub campaign_id_dimension_value: Option<DimensionValue>,
    pub primary: Option<bool>,
    pub placement_group_id: Option<String>,
    pub create_info: Option<LastModifiedInfo>,
    pub site_id: Option<String>,
    pub site_id_dimension_value: Option<DimensionValue>,
    pub compatibility: Option<String>,
    pub placement_strategy_id: Option<String>,
    pub placement_group_id_dimension_value: Option<DimensionValue>,
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub lookback_configuration: Option<LookbackConfiguration>,
    pub last_modified_info: Option<LastModifiedInfo>,
    pub pricing_schedule: Option<PricingSchedule>,
}

Contains properties of a placement.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

generatetags placements (none)

insert placements (request|response)

update placements (request|response)

list placements (none)

get placements (response)

patch placements (request|response)

Fields
comment: Option<String>

Comments for this placement.

campaign_id: Option<String>

Campaign ID of this placement. This field is a required field on insertion.

payment_source: Option<String>

Payment source for this placement. This is a required field that is read-only after insertion.

directory_site_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the directory site. This is a read-only, auto-generated field.

advertiser_id: Option<String>

Advertiser ID of this placement. This field can be left blank.

key_name: Option<String>

Key name of this placement. This is a read-only, auto-generated field.

directory_site_id: Option<String>

Directory site ID of this placement. On insert, you must set either this field or the siteId field to specify the site associated with this placement. This is a required field that is read-only after insertion.

advertiser_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the advertiser. This is a read-only, auto-generated field.

content_category_id: Option<String>

ID of the content category assigned to this placement.

size: Option<Size>

Size associated with this placement. When inserting or updating a placement, only the size ID field is used. This field is required on insertion.

archived: Option<bool>

Whether this placement is archived.

publisher_update_info: Option<LastModifiedInfo>

Information about the last publisher update. This is a read-only field.

tag_formats: Option<Vec<String>>

Tag formats to generate for this placement. This field is required on insertion. Acceptable values are: - "PLACEMENT_TAG_STANDARD" - "PLACEMENT_TAG_IFRAME_JAVASCRIPT" - "PLACEMENT_TAG_IFRAME_ILAYER" - "PLACEMENT_TAG_INTERNAL_REDIRECT" - "PLACEMENT_TAG_JAVASCRIPT" - "PLACEMENT_TAG_INTERSTITIAL_IFRAME_JAVASCRIPT" - "PLACEMENT_TAG_INTERSTITIAL_INTERNAL_REDIRECT" - "PLACEMENT_TAG_INTERSTITIAL_JAVASCRIPT" - "PLACEMENT_TAG_CLICK_COMMANDS" - "PLACEMENT_TAG_INSTREAM_VIDEO_PREFETCH" - "PLACEMENT_TAG_TRACKING" - "PLACEMENT_TAG_TRACKING_IFRAME" - "PLACEMENT_TAG_TRACKING_JAVASCRIPT"

payment_approved: Option<bool>

Whether payment was approved for this placement. This is a read-only field relevant only to publisher-paid placements.

tag_setting: Option<TagSetting>

Tag settings for this placement.

id: Option<String>

ID of this placement. This is a read-only, auto-generated field.

external_id: Option<String>

External ID for this placement.

id_dimension_value: Option<DimensionValue>

Dimension value for the ID of this placement. This is a read-only, auto-generated field.

account_id: Option<String>

Account ID of this placement. This field can be left blank.

ssl_required: Option<bool>

Whether creatives assigned to this placement must be SSL-compliant.

status: Option<String>

Third-party placement status.

campaign_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the campaign. This is a read-only, auto-generated field.

primary: Option<bool>

Whether this placement is the primary placement of a roadblock (placement group). You cannot change this field from true to false. Setting this field to true will automatically set the primary field on the original primary placement of the roadblock to false, and it will automatically set the roadblock's primaryPlacementId field to the ID of this placement.

placement_group_id: Option<String>

ID of this placement's group, if applicable.

create_info: Option<LastModifiedInfo>

Information about the creation of this placement. This is a read-only field.

site_id: Option<String>

Site ID associated with this placement. On insert, you must set either this field or the directorySiteId field to specify the site associated with this placement. This is a required field that is read-only after insertion.

site_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the site. This is a read-only, auto-generated field.

compatibility: Option<String>

Placement compatibility. DISPLAY and DISPLAY_INTERSTITIAL refer to rendering on desktop, on mobile devices or in mobile apps for regular or interstitial ads respectively. APP and APP_INTERSTITIAL are no longer allowed for new placement insertions. Instead, use DISPLAY or DISPLAY_INTERSTITIAL. IN_STREAM_VIDEO refers to rendering in in-stream video ads developed with the VAST standard. This field is required on insertion.

placement_strategy_id: Option<String>

ID of the placement strategy assigned to this placement.

placement_group_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the placement group. This is a read-only, auto-generated field.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#placement".

subaccount_id: Option<String>

Subaccount ID of this placement. This field can be left blank.

name: Option<String>

Name of this placement.This is a required field and must be less than 256 characters long.

lookback_configuration: Option<LookbackConfiguration>

Lookback window settings for this placement.

last_modified_info: Option<LastModifiedInfo>

Information about the most recent modification of this placement. This is a read-only field.

pricing_schedule: Option<PricingSchedule>

Pricing schedule of this placement. This field is required on insertion, specifically subfields startDate, endDate and pricingType.