Struct google_dfareporting2d5::TechnologyTargeting

pub struct TechnologyTargeting {
    pub browsers: Option<Vec<Browser>>,
    pub operating_systems: Option<Vec<OperatingSystem>>,
    pub mobile_carriers: Option<Vec<MobileCarrier>>,
    pub connection_types: Option<Vec<ConnectionType>>,
    pub platform_types: Option<Vec<PlatformType>>,
    pub operating_system_versions: Option<Vec<OperatingSystemVersion>>,
}

Technology Targeting.

This type is not used in any activity, and only used as part of another schema.

Fields
browsers: Option<Vec<Browser>>

Browsers that this ad targets. For each browser either set browserVersionId or dartId along with the version numbers. If both are specified, only browserVersionId will be used. The other fields are populated automatically when the ad is inserted or updated.

operating_systems: Option<Vec<OperatingSystem>>

Operating systems that this ad targets. To target specific versions, use operatingSystemVersions. For each operating system only dartId is required. The other fields are populated automatically when the ad is inserted or updated. If targeting an operating system, do not set targeting for operating system versions for the same operating system.

mobile_carriers: Option<Vec<MobileCarrier>>

Mobile carriers that this ad targets. For each mobile carrier only id is required, and the other fields are populated automatically when the ad is inserted or updated. If targeting a mobile carrier, do not set targeting for any zip codes.

connection_types: Option<Vec<ConnectionType>>

Connection types that this ad targets. For each connection type only id is required. The other fields are populated automatically when the ad is inserted or updated.

platform_types: Option<Vec<PlatformType>>

Platform types that this ad targets. For example, desktop, mobile, or tablet. For each platform type, only id is required, and the other fields are populated automatically when the ad is inserted or updated.

operating_system_versions: Option<Vec<OperatingSystemVersion>>

Operating system versions that this ad targets. To target all versions, use operatingSystems. For each operating system version, only id is required. The other fields are populated automatically when the ad is inserted or updated. If targeting an operating system version, do not set targeting for the corresponding operating system in operatingSystems.